requirements.txt
absl-py==0.12.0
antlr4-python3-runtime==4.9.3
apache-beam==2.34.0
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
avro-python3==1.9.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
cachetools==4.2.4
certifi==2023.5.7
charset-normalizer==3.1.0
clang==5.0
contourpy==1.0.7
crcmod==1.7
cycler==0.11.0
decorator==5.1.1
dill==0.3.1.1
docopt==0.6.2
executing==1.2.0
fastavro==1.7.4
filelock==3.12.0
flatbuffers==1.12
fonttools==4.39.4
future==0.18.3
gast==0.4.0
gdown==4.5.4
gin-config==0.5.0
google-api-core==1.34.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery-storage==1.1.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.48.2
h5py==3.1.0
hdfs==2.7.0
httplib2==0.19.1
hydra-core==1.3.2
idna==3.4
imageio==2.29.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipython==8.13.2
jedi==0.18.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mediapy==1.0.3
natsort==8.1.0
networkx==3.1
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
orjson==3.9.0
packaging==23.1
parameterized==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==5.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydot==1.4.2
Pygments==2.15.1
pymongo==3.13.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.1
scipy==1.10.1
six==1.15.0
soupsieve==2.4.1
stack-data==0.6.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-addons==0.15.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.6.0
tensorflow-metadata==1.13.1
termcolor==1.1.0
tifffile==2023.4.12
tqdm==4.64.1
traitlets==5.9.0
typeguard==2.13.3
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.12.1
zipp==3.15.0
--extra-index-url https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117